requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
langchain-text-splitters
faiss-cpu
sentence-transformers
python-dotenv
fastapi
uvicorn
